requirements.txt
# oqs library available locally
cryptography
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
eventlet==0.33.3 